keymap update

+++ .setup_gtags.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 echo "setup gtags:"
-if [ -e $PWD/plugin_script/gtags.vim ] || [ -e $PWD/plugin_script/gtags-cscope.vim ]; then
+if [ -e $PWD/plugin_script/gtags.vim ] && [ -e $PWD/plugin_script/gtags-cscope.vim ]; then
     mkdir -pv ~/.vim/plugin
     if [ -e ~/.vim/plugin/gtags.vim ]; then
         mv -v ~/.vim/plugin/gtags.vim ~/.vim/plugin/gtags.vim.org
